backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
